// src/app/_core/models/User.ts
export interface User {
  lastName: string;
  firstName: string;
  gender: string;
  profileCompleted: boolean;
  profilePhotoURL: string;
  age: Number;
  height: Number;
  weight: Number;
  eyeColor: string;
  hairColor: string;
  tattoo: string;
  personalityTraits: string;
  hobbies: string;
  sports: string;
  music: string;
  instagramURL: string;
  facebookURL: string;
  twitterURL: string;
  matches: Number;
  matchesAdded: string;
  email: string;
  password: string;
}